SPC-674 fix race condition in requisitions

+++ opennms-provision/opennms-provision-persistence/src/main/java/org/opennms/netmgt/provision/persist/FusedForeignSourceRepository.java
@@ -34,6 +34,7 @@ import java.util.Date;
 import java.util.List;
 import java.util.Set;
 
+import org.opennms.core.utils.LogUtils;
 import org.opennms.netmgt.provision.persist.foreignsource.ForeignSource;
 import org.opennms.netmgt.provision.persist.requisition.Requisition;
 import org.springframework.beans.factory.InitializingBean;
@@ -217,8 +218,8 @@ public class FusedForeignSourceRepository extends AbstractForeignSourceRepositor
      * @throws org.opennms.netmgt.provision.persist.ForeignSourceRepositoryException if any.
      */
     public synchronized void save(final Requisition requisition) throws ForeignSourceRepositoryException {
-        cleanUpSnapshots(requisition);
         m_deployedForeignSourceRepository.save(requisition);
+        cleanUpSnapshots(requisition);
     }
 
     private void cleanUpSnapshots(final Requisition requisition) {
@@ -227,30 +228,35 @@ public class FusedForeignSourceRepository extends AbstractForeignSourceRepositor
 
         final List<File> pendingSnapshots = RequisitionFileUtils.findSnapshots(m_pendingForeignSourceRepository, foreignSource);
 
-        /* determine whether to delete the pending requisition */
-        boolean deletePendingRequisition = true;
-        if (pendingSnapshots.size() > 0) {
-            for (final File pendingSnapshotFile : pendingSnapshots) {
-                if (isNewer(pendingSnapshotFile, pendingDate)) {
-                    // the pending file is newer than an in-process snapshot, don't delete it
-                    deletePendingRequisition = false;
-                    break;
+        if (pendingDate != null) {
+            /* determine whether to delete the pending requisition */
+            boolean deletePendingRequisition = true;
+            if (pendingSnapshots.size() > 0) {
+                for (final File pendingSnapshotFile : pendingSnapshots) {
+                    if (isNewer(pendingSnapshotFile, pendingDate)) {
+                        // the pending file is newer than an in-process snapshot, don't delete it
+                        deletePendingRequisition = false;
+                        break;
+                    }
                 }
             }
-        }
-        if (deletePendingRequisition) {
-            m_pendingForeignSourceRepository.delete(requisition);
+            if (deletePendingRequisition) {
+                m_pendingForeignSourceRepository.delete(requisition);
+            }
         }
 
         /* determine whether this requisition was imported from a snapshot, and if so, delete its snapshot file */
         RequisitionFileUtils.deleteResourceIfSnapshot(requisition);
+
+        final Date deployedDate = m_deployedForeignSourceRepository.getRequisitionDate(foreignSource);
+        LogUtils.debugf(this, "deployed date = %s", deployedDate);
+        if (deployedDate != null) {
+            RequisitionFileUtils.deleteSnapshotsOlderThan(getPendingForeignSourceRepository(), foreignSource, deployedDate);
+        }
     }
 
     private boolean isNewer(final File snap, final Date date) {
-        final String name = snap.getName();
-        final String timestamp = name.substring(name.lastIndexOf(".") + 1);
-        final Date snapshotDate = new Date(Long.valueOf(timestamp));
-        return date.after(snapshotDate);
+        return RequisitionFileUtils.isNewer(snap, date);
     }
 
     @Override
